requirements.txt
discord.py
ollama
pyyaml
requests
pollinations
yt-dlp
TTS
